pyproject.toml
[project]
name = "wordwill"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.13.0",
    "pillow>=12.0.0",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.2.1",
    "socksio>=1.0.0",
    "pyperclip>=1.9.0",
]
